# target1-app/openapi.yaml
openapi: 3.0.3

info:
  title: Access Control Test Target API
  version: 1.0.0
  description: Test application for automated access-control testing.

servers:
  - url: http://127.0.0.1:5000

paths:

  /:
    get:
      summary: Home
      responses:
        "200":
          description: Application is running

  /api/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials

  /api/products:
    get:
      summary: List products
      responses:
        "200":
          description: Product list

  /api/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
        "401":
          description: Authentication required

  /api/users/{user_id}:
    get:
      summary: Get user profile
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
        "403":
          description: Access denied
        "404":
          description: User not found

    patch:
      summary: Update user profile
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
      responses:
        "200":
          description: Profile updated
        "403":
          description: Access denied
        "404":
          description: User not found

  /api/orders/{order_id}:
    get:
      summary: Get order
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
        "403":
          description: Access denied
        "404":
          description: Order not found

  /api/admin/users:
    get:
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
        "403":
          description: Admin access required

  /api/admin/users/{user_id}:
    get:
      summary: Get any user as administrator
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
        "403":
          description: Admin access required
        "404":
          description: User not found

    delete:
      summary: Delete user as administrator
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
        "403":
          description: Admin access required
        "404":
          description: User not found

  /api/admin/stats:
    get:
      summary: Administrator statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Statistics
        "403":
          description: Admin access required

components:

  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT